(modified): dekstop view cleaned and ordered
